Modularize sensor.py: split into domain-specific modules and shared parsers
- Create parsers.py with shared parsing/conversion functions (no HA dependency)
- Create sensors/ package with domain-specific modules:
  - space_weather.py (672 lines) - Geomagnetic, K-index, Aurora, Solar Radiation
  - hurricanes.py (195 lines) - Hurricane alerts and activity
  - surf.py (239 lines) - Rip current, surf height, water temperature
  - weather_observations.py (478 lines) - Weather observation base + 9 sensors
  - forecasts.py (308 lines) - Extended and hourly forecast
  - alerts.py (110 lines) - NWS active alerts
  - weather_extra.py (412 lines) - Cloud cover, radar, forecast discussion
- Reduce sensor.py to 172-line setup entry point
- All files under 1000 lines, zero linting errors

Agent-Logs-Url: https://github.com/turbo5000c/dev-noaa_it_all/sessions/e7454924-70b9-4043-b7ed-7c3ad887fc15
